Fix accounts associted files issues arousing while examining its workflow

+++ PointOfSaleSystem.Data/Inventory/OtherTax.cs
@@ -9,7 +9,7 @@ namespace PointOfSaleSystem.Data.Inventory
         [Key]
         public int OtherTaxID { get; set; }
         public string OtherTaxName { get; set; } = "";
-        public int PerRate { get; set; }
+        public double PerRate { get; set; }
 
         [ForeignKey("VATLiabSubAccountID")]
         public SubAccount SubAccount { get; set; }
