+++ PointOfSaleSystem.Data/Inventory/VatType.cs
@@ -9,7 +9,7 @@ namespace PointOfSaleSystem.Data.Inventory
         [Key]
         public int VATTypeID { get; set; }
         public string VATTypeName { get; set; } = "";
-        public int PerRate { get; set; }
+        public double PerRate { get; set; }
 
         [ForeignKey("VATLiabSubAccountID")]
         public SubAccount SubAccount { get; set; } = new SubAccount();
